requirements.txt
-e .
aws_cdk.aws.ec2
aws_cdk.aws.ecs
aws_cdk.aws_ecs_pattern